Update_028
Change the shortcut to V uniquely in Paint mode.
One only function on one line: "Add Shaderless Painting Camera"
Return "Create brush maker Scene" in MACRO box in First position.
After Covert in Mesh Mask & mesh Mask Invertion, you return on the  active canvas in Paint mode.
Later i will add the same mat as the canvas too. But so, you can use the V shortcut in paint mode on the canvas in priority.
Main rule: to be on the main canvas to work. if you must work on a mask, you choose after.

But a problem in name of masks and inverted masks in automatic naming! But not a problem to draw or use it now.